Add metadata signing.
This patch adds metadata signing. It can be enabled globally by setting
the 'metadata.sign.enable' global option, or per SP/IdP by setting
the same option in the metadata of the SP/IdP.

To sign the metadata, both a private key and a certificate is used. The
following configuration options determine which key and certificate files
will be used:

- per SP/IdP metadata:
  'metadata.sign.privatekey' & 'metadata.sign.certificate'
- globally:
  'metadata.sign.privatekey' & 'metadata.sign.certificate'
- per SP/IdP metadata: (fallback)
  'privatekey' & 'certificate'