IdP discovery: Add support for saving the users choice of IdP.
A config option is added which enables this: 'idpdisco.enableremember'
A checkbox allowing the user to save his choice of IdP appears on the IdP
selection page if this option is enabled.

This patch breaks backwards-compatibility for the selectidp-links page. Also,
the selectidp-dropdown page must be modified to show the checkboxk.

An additional page (https://.../simplesaml/cleardiscochoices.php) has been
added to make it possible to delete the current IdP choices.

There are also some dictionary updates which are not part of this patch:
frontpage.php:
+	'link_cleardiscochoices' => array (
+		'en' => 'Delete my choices of IdP in the IdP discovery services',
+		'no' => 'Slett mitt valg av IdP i IdP discovery tjenestene',
+	),
disco.php:
+	'remember' => array (
+		'en' => 'Remember my choice',
+		'no' => 'Husk mitt valg',
+	),