statistics:logcleaner: Fix insecure open of output file.
The logcleaner script by default creates the output file in /tmp.
Another user with access to the machine can create a symbolic link with
the same name as the output file. This will make the user running the
logcleaner script overwrite another file on the system.

This patch fixes this by making sure that we don't open existing files.

Thanks to Thijs Kinkhorst <thijs@uvt.nl> for reporting this bug.